// Client.Web/src/interfaces/DetalleInscripcion.ts
export interface DetalleInscripcion {
  nombresEstudiante: string;
  apellidosEstudiante: string;
  cedulaEstudiante: string;
  lugarNacimientoEstudiante: string;
  fechaNacimientoEstudiante: string;
  estadoPaisEstudiante: string;
  generoEstudiante: string;
  numHijo: number;
  correoEstudiante: string;
  condicionEspecialEstudiante: string;
  plantelProcedenciaEstudiante: string;
  direccionEstudiante: string;
  poseeCanaima: boolean;
  serialCanaima: string;
  tieneBeca: boolean;
  enteOtorgaBeca: string;
  seccion: string;
  tallaCamisa: string;
  tallaPantalon: string;
  numZapato: number;
  peso: number;
  nombreRepresentante: string;
  apellidoRepresentante: string;
  cedulaRepresentante: string;
  nivelEstudioRepresentante: string;
  profesionRepresentante: string;
  direccionRepresentante: string;
  telfHabitacionRepresentante: string;
  telfCelularRepresentante: string;
  telfFamiliarRepresentante: string;
  nombrePadre: string;
  apellidoPadre: string;
  cedulaPadre: string;
  nivelEstudioPadre: string;
  profesionPadre: string;
  direccionPadre: string;
  telfHabitacionPadre: string;
  telfCelularPadre: string;
  telfFamiliarPadre: string;
  nombreMadre: string;
  apellidoMadre: string;
  cedulaMadre: string;
  nivelEstudioMadre: string;
  profesionMadre: string;
  direccionMadre: string;
  telfHabitacionMadre: string;
  telfCelularMadre: string;
  telfFamiliarMadre: string;
  boletaPromocion: boolean;
  partidaNacimiento: null;
  copiaCedula: boolean;
  foto: boolean;
  notasCertificadas: boolean;
  copiaCedulaRepre: boolean;
  fotoRepre: boolean;
  autorizacionRepre: boolean;
}